usr: refactor process server (proc) for Zig error handling, OOP lists, and native C constants
- Refactored `usr/server/proc/` to use structured namespaces from `prog` and `task` modules (such as `prog.stdlib`, `prog.errno`, `prog.termios`, `task.prex`), eliminating monolithic `c` FFI references.
- Refactored internal functions across `proc_pid.zig`, `proc_fork.zig`, `proc_exit.zig`, `proc_sig.zig`, and `proc_tty.zig` to return native Zig error unions (`!void`, `!task.prex.pid_t`) and propagate errors using `try`.
- Implemented `toCError` and `catchToCError` in `ffi.zig` to map Zig error sets to POSIX errno integers at top-level message handlers.
- Replaced manual cleanup flags with native `errdefer prog.stdlib.free(...)` in `sys_fork` and `sys_setsid`.
- Replaced manual linked-list inline calls with an object-oriented intrusive `List` API matching `sys/lib/list.zig`.
- Implemented `allocZeroed` bounded slice helper in `ffi.zig`.
- Fixed `_IOC` macro in `include/sys/ioctl.h` with explicit `(u_long)` casts to ensure ioctl constants translate cleanly into unsigned integers during `@cImport`.
- Added `zig_from_c_refactor.md` documenting best practices and architectural guidelines for C-to-Zig refactoring.
- Verified all 16 target configurations boot successfully.

+++ sdk/include/sys/ioctl.h
@@ -56,7 +56,7 @@
 #define IOC_IVAL 0x10000000 /* input argument is immediate value */
 #define IOC_OVAL 0x08000000 /* return value as output */
 
-#define _IOC(inout, group, num, len) (u_long)(inout | ((len & IOCPARM_MASK) << 16) | ((group) << 8) | (num))
+#define _IOC(inout, group, num, len) ((u_long)(inout) | ((u_long)((len) & IOCPARM_MASK) << 16) | ((u_long)(group) << 8) | (u_long)(num))
 #define _IO(g, n) _IOC(IOC_VOID, (g), (n), 0)
 #define _IOR(g, n, t) _IOC(IOC_OUT, (g), (n), sizeof(t))
 #define _IOW(g, n, t) _IOC(IOC_IN, (g), (n), sizeof(t))
